Implement common stats row element
